requirements.txt
asgiref==3.5.2
boto3==1.24.84
botocore==1.27.84
Django==3.2.15
django-dotenv==1.4.2
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
pillow==9.2.0
djangorestframework
markdown
django-filter
django-cors-headers
